ours/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
contourpy==1.1.1
cycler==0.12.1
fonttools==4.56.0
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.10.1
six==1.17.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
typing_extensions==4.13.0
zipp==3.20.2
